+++ src/ExplorePackages.Worker.Logic/MessageProcessors/IBatchMessageProcessor.cs
@@ -5,6 +5,6 @@ namespace Knapcode.ExplorePackages.Worker
 {
     public interface IBatchMessageProcessor<T>
     {
-        Task<BatchMessageProcessorResult<T>> ProcessAsync(IReadOnlyList<T> messages, int dequeueCount);
+        Task<BatchMessageProcessorResult<T>> ProcessAsync(IReadOnlyList<T> messages, long dequeueCount);
     }
 }
